Se cambia el aler para eliminar post

+++ src/lib/Components/dashboard.js
@@ -1,7 +1,7 @@
 
 import { onNavigate } from "../../main.js";
 import { app } from "../Firebase.js";
-import { submitPost, logOut, getAllPosts, deletePost, currentUserInfo, getTask,updateTask } from "../index.js";
+import { submitPost, logOut, getAllPosts, deletePost, currentUserInfo, getTask, updateTask } from "../index.js";
 
 
 export const login = () => {
@@ -80,7 +80,7 @@ export const login = () => {
         deleteIcon.setAttribute('src','/images/delete.png');
         deleteIcon.className = 'delete-icon';
         deleteIcon.setAttribute('data-id', post.id);//el data-id es algo que ya trae firebase
-  
+
         deleteIcon.onclick = deletePostListener;
         likePost.setAttribute('src', '/images/1erlike.png');
         likePost.className = 'primer-like';
@@ -115,45 +115,39 @@ export const login = () => {
     });
   };
 //listener de onclick editarPost
-  const editPostListener = async(event) => {
-    const doc = await getTask(event.target.dataset.id);
-    const task = doc.data();
-    let editStatus = false;
-    let newPost = task.postText;
-    divLogin.querySelector('#input-post').value= newPost
-    getTask(newPost)
-    .then((response) => {
-    console.log(response);
-    document.querySelector('#modal-background-post').style.display = 'flex';
-    document.querySelector('#modal-content-post').style.display = 'block';
-    document.body.style.overflow = 'hidden';
-    document.querySelector('#input-post').focus();
-    document.querySelector('#btn-post').disabled = false; // boton publicar activo
-    editStatus = true;
-    })
-    if (editStatus===true) {
-      updateTask(doc,{newPost});
-      console.log('updating',updateTask);
-    }
-    
-
-  };
+const editPostListener = async(event) => {
+  const doc = await getTask(event.target.dataset.id);
+  const task = doc.data();
+  let editStatus = false;
+  let newPost = task.postText;
+  divLogin.querySelector('#input-post').value= newPost
+  getTask(newPost)
+  .then((response) => {
+  console.log(response);
+  document.querySelector('#modal-background-post').style.display = 'flex';
+  document.querySelector('#modal-content-post').style.display = 'block';
+  document.body.style.overflow = 'hidden';
+  document.querySelector('#input-post').focus();
+  document.querySelector('#btn-post').disabled = false; // boton publicar activo
+  editStatus = true;
+  })
+  if (editStatus===true) {
+    updateTask(doc,{newPost});
+    console.log('updating',updateTask);
+  }; 
+};
+  
 
 //listener del onclick detelePost
   const deletePostListener = (event) => {//event por default
     const postId = event.target.dataset.id;//sacamos del target el id
-    console.log('delete clicked', postId);
-    //  if(postDatas.uid === currentUserInfo().uid){
-
-    deletePost(postId)
-    .then((response) => {
-      console.log(response);
-      alert(' Comentario borrado');
+      let opcion = confirm('Desea eliminar el comentario?');
+      if(opcion === false){}
+      else {
+        deletePost(postId);
+      };
       refreshPosts();
-    })
-    .catch(error => {console.log(error);});
-  };
-
+    };
 
   //aqui se manda llamar el getDocs al cargar la pagina en Dashboard
   refreshPosts();
